Implemented a GET endpoint for service hierarchy health status.
Closes BATAPI-86

+++ sonar-core/Prometheus/ResultData.cs
@@ -7,5 +7,5 @@ namespace Cms.BatCave.Sonar.Prometheus;
 public record ResultData(
   [property:JsonPropertyName("metric")]
   IImmutableDictionary<String, String> Labels,
-  (Int64 Timestamp, String Value)? Value,
-  IImmutableList<(Int64 Timestamp, String Value)>? Values);
+  (Decimal Timestamp, String Value)? Value,
+  IImmutableList<(Decimal Timestamp, String Value)>? Values);
